Added vault common
