+++ programs/vault/src/state/mod.rs
@@ -1,7 +1,9 @@
 pub mod constants;
 pub mod fee;
-pub mod vault_config;
+pub mod vault;
+pub mod vault_common;
 
 pub use constants::*;
 pub use fee::*;
-pub use vault_config::*;
+pub use vault::*;
+pub use vault_common::*;
